+++ deno/deps.ts
@@ -1,3 +1,3 @@
-export type { Sql } from "npm:postgres@^3.4.3";
+export type { Sql } from "npm:postgres@^3.4.4";
 export { parseArgs } from "https://deno.land/std@0.211.0/cli/mod.ts";
 export { extname } from "https://deno.land/std@0.211.0/path/extname.ts";