+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file documents notable changes to the project
 **new**
 
 - add the second argument to `config.getUsage(struct, currentValue)` so that is also documents the current configuration value after the default one
+- `ConfigurationOptions#readTextField` accepts URLs and strings
 
 **fixes**